# Better performance with Activator.CreateInstance