@304NotModified 2018-07-08
Also backed here: https://stackoverflow.com/a/9789971/201303

any idea why it's faster?