@tangdf 2018-07-09
Expressions are the best performance, but typically, we need to add a dictionary to cache delegates, so the performance is very close.

The reason for using `Activatator.Createinstance` is that it is very lean.

I'm so sorry, i have no idea  why it's faster!


```c#
    public class CreateInstanceBenchamrker
    {
        private static readonly Type s_type = typeof(TestClass);

        [Benchmark]
        public object CreateInstance_By_FactoryHelper() {
            return FactoryHelper.CreateInstance(s_type);
        }

        [Benchmark]
        public object CreateInstance_By_Activator() {
            return Activator.CreateInstance(s_type);
        }


        [Benchmark]
        public object CreateInstance_By_Expression() {
            return ExpressionFactoryHelper.CreateInstance(s_type);
        }
    }

    public static class ExpressionFactoryHelper
    {
        private static readonly ConcurrentDictionary<Type, Func<object>> s_cache = new ConcurrentDictionary<Type, Func<object>>();

        internal static object CreateInstance(Type t) {

            return s_cache.GetOrAdd(t, CreateFunc).Invoke();
        }

        private static Func<object> CreateFunc(Type t) {

            ConstructorInfo constructor = t.GetConstructor(Array.Empty<Type>());
            if (constructor != null) {
                var newExpression = Expression.New(constructor);
                var convertExpression = Expression.Convert(newExpression,typeof(object));
                return Expression.Lambda<Func<object>>(convertExpression).Compile();
            }
            else {
                return () => throw new InvalidOperationException(
                    $"Cannot access the constructor of type: {t.FullName}. Is the required permission granted?");
            }
        }
    }
```

```
|                          Method |      Mean |     Error |    StdDev |
|-------------------------------- |----------:|----------:|----------:|
| CreateInstance_By_FactoryHelper | 209.61 ns | 0.9511 ns | 0.8897 ns |
|     CreateInstance_By_Activator |  54.38 ns | 0.1308 ns | 0.1224 ns |
|    CreateInstance_By_Expression |  40.09 ns | 0.2844 ns | 0.2521 ns |